fix: rax register to be signed 64 bit

+++ twilight_kernel/src/arch/x86_64/idt.rs
@@ -18,7 +18,7 @@ pub struct Registers {
     pub rsi: usize,
     pub rdx: usize,
     pub rcx: usize,
-    pub rax: usize,
+    pub rax: i64,
 }
 
 // Translate IRQ into system interrupt
